Moved VectorBase to it's dedicated folder Seperated concat and cross product as partial classes from Vector3 Updated Vector3 to use the new Memory management system
