+++ DataScience/Core/VectorBase/VectorBase.cs
