+++ DataScience/DataScience.csproj
@@ -20,7 +20,6 @@
 
   <ItemGroup>
     <Folder Include="Astrophysics\" />
-    <Folder Include="Core\VectorBase\" />
     <Folder Include="Tests\" />
   </ItemGroup>
